requirements.txt
pandas
numpy
scikit-learn
sentence-transformers
pyyaml
fastparquet